functional extension refactor
