+++ ivy/functional/ivy/extensions/statistical.py
@@ -1,0 +1,55 @@
+from typing import (
+    Optional,
+    Union,
+    Tuple,
+)
+import ivy
+from ivy.func_wrapper import (
+    handle_out_argument,
+    to_native_arrays_and_back,
+    handle_nestable,
+)
+from ivy.exceptions import handle_exceptions
+
+
+@to_native_arrays_and_back
+@handle_out_argument
+@handle_nestable
+@handle_exceptions
+def median(
+    input: ivy.Array,
+    /,
+    *,
+    axis: Optional[Union[Tuple[int], int]] = None,
+    keepdims: Optional[bool] = False,
+    out: Optional[ivy.Array] = None,
+) -> ivy.Array:
+    """Compute the median along the specified axis.
+
+    Parameters
+    ----------
+    input
+        Input array.
+    axis
+        Axis or axes along which the medians are computed. The default is to compute
+        the median along a flattened version of the array.
+    keepdims
+        If this is set to True, the axes which are reduced are left in the result
+        as dimensions with size one.
+    out
+        optional output array, for writing the result to.
+
+    Returns
+    -------
+    ret
+        The median of the array elements.
+
+    Functional Examples
+    -------------------
+    >>> a = ivy.array([[10, 7, 4], [3, 2, 1]])
+    >>> ivy.median(a)
+    3.5
+    >>> ivy.median(a, axis=0)
+    ivy.array([6.5, 4.5, 2.5])
+    """
+    return ivy.current_backend().median(input, axis=axis, keepdims=keepdims, out=out)
